backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
PyMySQL==1.1.1
passlib==1.7.4
python-dotenv==1.1.1
email-validator==2.2.0
qrcode==8.2
